Adds vtkOpenGLTextureImage class
Refactor 3D textures out of the shader computation.  Now there
can be multiple textures that are referenced in the shaders in
terms of the corresponding texture unit.

This simplifies shader computation and generalizes what can be done.

This should ultimately support compsiting, label map transfer functions,
and nonlinear transforms as vector fields images.